requirements.txt
# Student AI Travel Planner - Required Packages
# Install with: pip install -r requirements.txt

# Core web framework
streamlit>=1.28.0

# AI Provider API
google-generativeai>=0.3.0  # For Gemini API (FREE)

# Optional: For enhanced functionality
requests>=2.31.0            # For HTTP requests (usually included with other packages)
python-dotenv>=1.0.0        # For environment variable management (optional)

# Development and testing (optional)
pytest>=7.4.0              # For running tests
